requirements.txt
torch
torchvision
pillow
opencv-python
matplotlib
streamlit
numpy
